+++ samples/FleetProvisioning/src/main/java/fleetprovisioning/FleetProvisioningSample.java
@@ -174,7 +174,7 @@ public class FleetProvisioningSample {
             builder.withConnectionEventCallbacks(callbacks)
                 .withClientId(cmdData.input_clientId)
                 .withEndpoint(cmdData.input_endpoint)
-                .withPort((short)cmdData.input_port)
+                .withPort(cmdData.input_port)
                 .withCleanSession(true)
                 .withProtocolOperationTimeoutMs(60000);
             connection = builder.build();
